+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add new `Card` Mission Control `Job` kinds.
+
 ## [0.4.2]
 
 - Fix `Folder` and `FoldingCard` Mission Control `Job` kinds being listed as available kinds 
@@ -83,8 +85,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add finishings specification
 - Add jobs and parts specification
 
-[Unreleased]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.4.2..HEAD
-[0.4.2]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.4.0..v0.4.2
+[Unreleased]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.4.1..HEAD
 [0.4.1]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.4.0..v0.4.1
 [0.4.0]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.3.0..v0.4.0
 [0.3.0]: https://github.com/zaikio/zaikio-mission_control-ruby/compare/v0.2.12..v0.3.0